+++ src/jvmMain/kotlin/jp/kaleidot725/adbpad/domain/model/command/NormalCommand.kt
@@ -1,7 +1,7 @@
 package jp.kaleidot725.adbpad.domain.model.command
 
 import com.malinskiy.adam.request.shell.v1.ShellCommandRequest
-import jp.kaleidot725.adbpad.domain.model.Language
+import jp.kaleidot725.adbpad.domain.model.language.Language
 
 interface NormalCommand {
     val title: String
