+++ src/jvmMain/kotlin/jp/kaleidot725/adbpad/domain/model/setting/Appearance.kt
@@ -1,6 +1,6 @@
 package jp.kaleidot725.adbpad.domain.model.setting
 
-import jp.kaleidot725.adbpad.domain.model.Language
+import jp.kaleidot725.adbpad.domain.model.language.Language
 
 enum class Appearance(val value: String) {
     DARK(Language.DARK),
