+++ src/jvmMain/kotlin/jp/kaleidot725/adbpad/view/component/device/DeviceSelector.kt
@@ -16,9 +16,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import jp.kaleidot725.adbpad.domain.model.Language
 import jp.kaleidot725.adbpad.domain.model.device.Device
 import jp.kaleidot725.adbpad.domain.model.device.DeviceState
+import jp.kaleidot725.adbpad.domain.model.language.Language
 
 @Composable
 fun DeviceSelector(
