+++ src/jvmMain/kotlin/jp/kaleidot725/adbpad/view/component/screenshot/ScreenshotViewer.kt
@@ -13,7 +13,7 @@ import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.loadImageBitmap
-import jp.kaleidot725.adbpad.domain.model.Language
+import jp.kaleidot725.adbpad.domain.model.language.Language
 import jp.kaleidot725.adbpad.domain.model.screenshot.ScreenshotPreview
 
 @Composable
