+++ src/jvmMain/kotlin/jp/kaleidot725/adbpad/view/component/setting/SettingHeader.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import jp.kaleidot725.adbpad.domain.model.Language
+import jp.kaleidot725.adbpad.domain.model.language.Language
 
 @Composable
 fun SettingHeader(
