+++ src/jvmMain/kotlin/jp/kaleidot725/adbpad/view/component/text/TextCommandList.kt
@@ -16,8 +16,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import jp.kaleidot725.adbpad.domain.model.Language
 import jp.kaleidot725.adbpad.domain.model.command.TextCommand
+import jp.kaleidot725.adbpad.domain.model.language.Language
 
 @Composable
 fun TextCommandList(
